% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.r
\name{distance}
\alias{distance}
\title{Calculate distance in psychological space}
\usage{
distance(w, attrs, anchor, r)
}
\arguments{
\item{w}{A vector of attribute weights.}

\item{attrs}{A matrix with ncol(attrs) = length(weights)}

\item{anchor}{An "exemplar" or anchor point to which attrs will be compared}

\item{r}{A positive value between that changes the metric form}
}
\value{
A vector of distances for each row of attrs
}
\description{
Take a vector of weights and an anchor point, and a set of attribute values
and return the distance between each point represented by the attribute
values and the anchor, adjusted by the metric form \code{r}.
}
