% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{ll_double}
\alias{ll_double}
\title{Calculate likelihood of x given double option data (N attrs)}
\usage{
ll_double(x, data, rp_func = rp_double, ...)
}
\arguments{
\item{x}{A named vector of parameters. Expects certain values as in
the description for the \code{rp_func}}

\item{data}{A data.frame compatible object with specific rows as described.}

\item{rp_func}{A function for calculating response probabilities for choosing
the left presented item.}

\item{...}{Other parameters passed to \code{rp_func}}
}
\value{
A single value (log-likelihood)
}
\description{
This function transforms parameters from the real number line to
the appropriate regions using \code{transform_pars_dbl}. It then uses a
sensitivity adjusted Luce Choice Axiom rule and distances from the
exemplar as from the single option models to calculate the likelihood
of the parameter vector x given the data.
}
\details{
The data object is assumed to be a data.frame compatible object with at
least columns for each attribute used in the \code{rp_func} and additionally
have a \code{response} column with T/F for whether the option selected was on the
left.
}
