% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{ll_optim}
\alias{ll_optim}
\title{Likelihood for single option, but minimises.}
\usage{
ll_optim(x, data)
}
\arguments{
\item{x}{A named vector of parameters. Expects certain values as in
description}

\item{data}{A data.frame compatible object with specific rows as described.}
}
\value{
negated likelihood
}
\description{
See description for ll_pmwg. This is a thin wrapper that negates the output
for use with the R built-in optim.
}
