% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{ll_pmwg}
\alias{ll_pmwg}
\title{Calculate likelihood of x given the data and return it}
\usage{
ll_pmwg(x, data, sample = FALSE)
}
\arguments{
\item{x}{A named vector of parameters. Expects certain values as in
description}

\item{data}{A data.frame compatible object with specific rows as described.}

\item{sample}{A boolean which specifies whether to return the likelihood, or
sample new responses.}
}
\value{
A single value (log-likelihood) or a copy of \code{data} with new
responses
}
\description{
This function transforms parameters from the real number line to
the appropriate regions using \code{transform_pars}. The using a variation
of the generalised context model for categorisation it calculates
the likelihood of the parameter vector x at leading to the accept/reject
responses to the prices and ratings in the data.
}
\details{
If \code{sample = TRUE} it will instead replace the accept column of the data with
values sampled from the model using the provided parameter values x

The \code{x} parameter vector is expected to be a named vector containing values
for \code{w} weight on the price attribute, \code{r} form of the distance metric,
\code{s} sensitivity to distances in attreibute space and \code{delta} a choice offset
relative to the anchor.

The data object is assumed to be a data.frame compatible object with at
least three columns, a \code{price_n} column with values for the price attribute
normalised to 0-1 and reversed so that 0 is the "worst" ie highest price and
1 is the best price. It should also have a rating_n column with ratings
normalised to 0-1 (worst-best) and finally an \code{accept} column with T/F for
whether the option was accepted.
}
