% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{normalise}
\alias{normalise}
\title{Normalises a column of values to the 0-1 range for estimation}
\usage{
normalise(val_col, val_rng = val_col, invert = FALSE)
}
\arguments{
\item{val_col}{A vector of raw attribute values.}

\item{val_rng}{A vector containning the range of possible values}

\item{invert}{If true return 1 - normalised value}
}
\value{
A vector of normalised values
}
\description{
Take a numeric vector of raw attribute values and a vector containing the
full range of possible values and normalise the values such that the
smallest real value is tranformed to 0 and the largest to 1. It is also
possible to invert the returned values for attributes such as price where
higher raw values should be lower.
}
