% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.r
\name{rp_cut}
\alias{rp_cut}
\title{Calculate response probabilities for 2xN using the cut_remain method}
\usage{
rp_cut(
  x,
  data,
  loc1_attrs = c("left_price", "left_memory"),
  loc2_attrs = c("right_price", "right_memory")
)
}
\arguments{
\item{x}{A named vector of parameters. Expects certain values as in
description}

\item{data}{A data.frame compatible object with specific rows as described.}

\item{loc1_attrs}{The column names containing normalised values for the
attributes (which has the weight value applied to it) in location 1}

\item{loc2_attrs}{The name of the column containing normalised values for the
attributes (which has the weight value applied to it) in location 2}
}
\description{
Used for both calculating likelihoods or sampling from a parameter estimate
This function calculates the probabilities of responses to a dataset
based on the provided parameters (in the \code{x} vector).
}
\details{
The \code{x} parameter vector is expected to be a named vector containing values
for \verb{\eqn{c_1..c_{N-1}}} where each \eqn{c_i} corresponds to the cut point
for each attribute \eqn{1..(N-1)}. The cut points are transformed to
attribute weights by cutting the remaining proportion of the 0-1 scale for
each new cut point, and the last weight being the remaining proportion.
The parameter vector \code{x} also contains \code{r} form of the distance metric and
\code{gamma} sensitivity to differences in attribute space.
}
