% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{rp_double}
\alias{rp_double}
\title{Calculate response probabilities for double option double attribute data}
\usage{
rp_double(
  x,
  data,
  loc1_attr1 = "left_price",
  loc1_attr2 = "left_memory",
  loc2_attr1 = "right_price",
  loc2_attr2 = "right_memory"
)
}
\arguments{
\item{x}{A named vector of parameters. Expects certain values as in
description}

\item{data}{A data.frame compatible object with specific rows as described.}

\item{loc1_attr1}{The name of the column containing normalised values for the
first attribute (which has the weight value applied to it) in location 1}

\item{loc1_attr2}{The name of the column containing normalised values for the
second attribute(which has the (1 - weight) values applied to it in
location 1}

\item{loc2_attr1}{The name of the column containing normalised values for the
first attribute (which has the weight value applied to it) in location 2}

\item{loc2_attr2}{The name of the column containing normalised values for the
second attribute(which has the (1 - weight) values applied to it in
location 2}
}
\description{
Used for both calculating likelihoods or sampling from a parameter estimate
This function calculates the probabilities of responses to a dataset
based on the provided parameters (in the \code{x} vector).
}
\details{
The \code{x} parameter vector is expected to be a named vector containing values
for \code{w} weight on the price attribute, \code{r} form of the distance metric and
\code{gamma} sensitivity to differences in attribute space.
}
