% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{rp_single}
\alias{rp_single}
\title{Calculate response probabilities for single option data}
\usage{
rp_single(x, data)
}
\arguments{
\item{x}{A named vector of parameters. Expects certain values as in
description}

\item{data}{A data.frame compatible object with specific rows as described.}
}
\description{
Used for both calculating likelihoods or sampling from a parameter estimate
This function calculates the probabilities of responses to a dataset
based on the provided parameters (in the \code{x} vector).
}
