% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.r
\name{rp_single_weight}
\alias{rp_single_weight}
\title{Calculate response probabilities for single option double attribute data}
\usage{
rp_single_weight(x, data, attrs = c("price_n", "rating_n"))
}
\arguments{
\item{x}{A named vector of parameters. Expects certain values as in
description}

\item{data}{A data.frame compatible object with specific rows as described.}

\item{attrs}{The column names containing normalised values for the
attributes (which has the weight value applied to it) in location 1}
}
\description{
Used for both calculating likelihoods or sampling from a parameter estimate
This function calculates the probabilities of responses to a dataset
based on the provided parameters (in the \code{x} vector).
}
