% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.r
\name{sample_single}
\alias{sample_single}
\title{Sample new responses from the parameter vector and single option data}
\usage{
sample_single(x, data)
}
\arguments{
\item{x}{A named vector of parameters. Expects certain values as in
description}

\item{data}{A data.frame compatible object with specific rows as described.}
}
\description{
Given a parameter vector \code{x} and a dtaset \code{data} generate response
probabilities for the rows in the data for single option trials.
}
\details{
The function \link{ll_single} explains the data format in more detail
}
