% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{transform_pars_dbl}
\alias{transform_pars_dbl}
\title{Transform the parameter vector (dbl option, dbl attr) for use in pmwg}
\usage{
transform_pars_dbl(x, fwd = TRUE)
}
\arguments{
\item{x}{The named vector of parameter estimates}

\item{fwd}{Move certain parameter to real number line or back}
}
\value{
transformed parameter vector x
}
\description{
See also the description for ll_double. This is a helper function that
transforms the parameters. To transform for pmwg the \code{r} and \code{gamma} values
are exponentiated in order to move to positive only and \code{w} is pnormed so
that values are between 0 and 1.
}
\details{
These operation are reversed when \code{fwd=TRUE}
}
