% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.r
\name{transform_pars_dir}
\alias{transform_pars_dir}
\title{Transform the parameter vector (2 x N) dirichlet method for use in pmwg}
\usage{
transform_pars_dir(x, fwd = TRUE)
}
\arguments{
\item{x}{The named vector of parameter estimates}

\item{fwd}{Move certain parameter to real number line or back}
}
\value{
transformed parameter vector x
}
\description{
See also the description for rp_dir This is a helper function that
transforms the parameters. To transform for pmwg all parameter values
are exponentiated in order to move to positive only.
}
\details{
These operation are reversed when \code{fwd=TRUE}
}
