% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.r
\name{distance}
\alias{distance}
\alias{distance.numeric}
\alias{distance.matrix}
\title{Calculate distances in psychological space}
\usage{
distance(w, attrs, anchor, r, ...)

\method{distance}{numeric}(w, attrs, anchor, r)

\method{distance}{matrix}(w, attrs, anchor, r)
}
\arguments{
\item{w}{A vector or matrix of attribute weights.}

\item{attrs}{A matrix with ncol(attrs) = length(weights)}

\item{anchor}{An "exemplar" or anchor point to which attrs will be compared}

\item{r}{A positive value between that changes the metric form}

\item{...}{Unused, for extensibility}
}
\value{
A vector of distances for each row of attrs
}
\description{
Take a collection of weights and an anchor point, and a set of attribute
values and return the distance between each point represented by the
attribute values and the anchor, adjusted by the metric form \code{r}.
}
