% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.r
\name{ll_double}
\alias{ll_double}
\title{Calculate likelihood of x given double option data (2 attrs)}
\usage{
ll_double(x, data, ...)
}
\arguments{
\item{x}{A named vector of parameters. Expects certain values as in
description}

\item{data}{A data.frame compatible object with specific rows as described.}

\item{...}{Other parameters passed to \code{rp_single}}
}
\value{
A single value (log-likelihood)
}
\description{
This function transforms parameters from the real number line to
the appropriate regions using \code{transform_pars_dbl}. It then uses a
sensitivity adjusted Luce Choice Axiom rule and distances from the
exemplar as from the single option models to calculate the likelihood
of the parameter vector x given the data.
}
\details{
The \code{x} parameter vector is expected to be a named vector containing values
for \code{w} weight on the price attribute, \code{r} form of the distance metric and
\code{gamma} sensitivity to differences in attribute space.

The data object is assumed to be a data.frame compatible object with at
least five columns, a \code{left_price} and a \code{right_price} column with values
for the price attribute normalised to 0-1 and reversed so that 0 is the
"worst" ie highest price and 1 is the best price.
It should also have a \code{left_memory} and a \code{right_memory} column with memory
values normalised to 0-1 (worst-best) and finally a \code{response} column with
T/F for whether the option selected was on the left.
}
